Готова регистрация с номера телефона в Telegram

+++ bot/handlers/user/cabinet_handlers.py
@@ -4,10 +4,12 @@ import os
 from PIL import Image
 from aiogram import Router, F
 from aiogram.types import Message, FSInputFile
-
+from sqlalchemy import select, exists
+from sqlalchemy.ext.asyncio import AsyncSession
 
 from api.iIkoCloud.enums import TypeRCI
 from api.iIkoCloud.iIkoCloud import IikoCloudAPI
+from bot.database.models.User import User
 from bot.keyboards.inline import sell_inline_kb
 from bot.mics.helpers.Config import Config
 from utils.main import generate_qr
@@ -18,27 +20,29 @@ iiko: IikoCloudAPI = IikoCloudAPI(api_login=Config.get('IIKOCLOUD_LOGIN'))
 
 # TODO: Проверить, что пользователь авторизован
 @router.message(F.text == 'Бонусная карта')
-async def profile_handler(msg: Message):
-    # TODO: Сделать генерацию и вывод QR кода
-
+async def profile_handler(msg: Message, session: AsyncSession):
     # TODO: Сделать dataclass с парамтерами профиля
     # TODO: Получить с БД номер телефона пользователя
-    profile_info = iiko.customer_info(
-        organization_id=Config.get('IIKOCLOUD_ORGANIZATIONS_IDS', 'list')[0],
-        identifier='79130478769',
-        type=TypeRCI.phone
-    )
-
     bot = msg.bot
 
-    generate_qr(text='79130478769', use_logo=False)
+    if await session.scalar(exists().where(User.user_id == msg.from_user.id).select()):
+        user = await session.scalars(select(User).where(User.user_id == msg.from_user.id))
+        user = user.first()
 
-    photo = FSInputFile('qr_code.png')
+        profile_info = iiko.customer_info(
+            organization_id=Config.get('IIKOCLOUD_ORGANIZATIONS_IDS', 'list')[0],
+            identifier=user.phone_number,
+            type=TypeRCI.phone
+        )
 
-    info = (f'Номер бонусной карты: {profile_info["phone"]}\n\n'
-            f'Бонусов: {round(profile_info["walletBalances"][0]["balance"])}')
+        generate_qr(text=user.phone_number, use_logo=False)
 
-    await bot.send_photo(chat_id=msg.chat.id, photo=photo, caption=info, reply_markup=sell_inline_kb())
+        photo = FSInputFile('qr_code.png')
 
-    os.remove('qr_code.png')
+        info = (f'Номер бонусной карты: {profile_info["phone"]}\n\n'
+                f'Бонусов: {round(profile_info["walletBalances"][0]["balance"])}')
+
+        await bot.send_photo(chat_id=msg.chat.id, photo=photo, caption=info, reply_markup=sell_inline_kb())
+
+        os.remove('qr_code.png')
